Pulled out vdu_control_panel and vdu_controller
