+++ vdu_controls/vdu_controls/vdu_controls_application.py
@@ -3,11 +3,9 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 from __future__ import annotations
 
-import glob
 import locale
 import math
 import pathlib
-import queue
 import random
 import select
 import signal
@@ -23,24 +21,24 @@ from dataclasses import dataclass
 from datetime import timedelta
 from functools import partial
 from importlib import import_module
-from typing import List, Tuple, Dict, Callable, Any, cast
+from typing import List, Tuple, Dict, Callable, cast
 
-from vdu_controls.config_ini import ConfIni, ConfOpt, VduControlsConfig, SUPPORTED_VCP_BY_CODE, VcpCapability, GeoLocation
+from vdu_controls.config_ini import ConfIni, ConfOpt, VduControlsConfig, VcpCapability, GeoLocation
 from vdu_controls.constants import *
 from vdu_controls.ddcutil_aggregator import DdcutilAggregator, VduStableId
 from vdu_controls.ddcutil_abstract import VcpOrigin, VcpValue, DdcutilDisplayNotFound, CONTINUOUS_TYPE, \
-    COMPLEX_NON_CONTINUOUS_TYPE, BRIGHTNESS_VCP_CODE, CONTRAST_VCP_CODE, SIMPLE_NON_CONTINUOUS_TYPE, DdcEventType, \
+    BRIGHTNESS_VCP_CODE, DdcEventType, \
     DdcutilServiceNotFound
 from vdu_controls.ddcutil_emulator import DdcutilEmulatorImpl
 from vdu_controls.ddcutil_laptop_panel import DdcutilPanelImpl
 from vdu_controls.greyscale import GreyScaleDialog
 from vdu_controls.help import HelpDialog
-from vdu_controls.icon_utils import ThemeType, polychrome_light_or_dark, SVG_LIGHT_THEME_COLOR, get_splash_image
+from vdu_controls.icon_utils import ThemeType, SVG_LIGHT_THEME_COLOR, get_splash_image
 from vdu_controls.icon_utils import create_image_from_svg_bytes, \
     create_icon_from_svg_bytes, create_icon_from_path, create_decorated_app_icon, StdPixmap, \
     is_dark_theme, si
 from vdu_controls.installer import install_as_desktop_application
-from vdu_controls.internationalization import tr, initialise_locale_translations, find_locale_specific_file, translate_option
+from vdu_controls.internationalization import tr, initialise_locale_translations, find_locale_specific_file
 from vdu_controls.logging import *
 from vdu_controls.misc import intV, zoned_now, proper_name, clamp
 from vdu_controls.preset import Preset, PresetScheduleStatus, PresetTransitionFlag
@@ -50,15 +48,17 @@ from vdu_controls.qt_imports import *
 from vdu_controls.release import release_notes
 from vdu_controls.scaling import native_font_height, npx
 from vdu_controls.settings_editor import SettingsDialog
-from vdu_controls.solar_calc import calc_solar_lux, create_elevation_map, SolarElevationKey, \
-    SolarElevationData, format_solar_elevation_abbreviation, parse_solar_elevation_ini_text, format_solar_elevation_ini_text
+from vdu_controls.solar_calc import calc_solar_lux, create_elevation_map
 from vdu_controls.svg import *
 from vdu_controls.unicode import *
 from vdu_controls.vdu_bulk_change import BulkChangeWorker, BulkChangeItem
+from vdu_controls.vdu_control_panel import VduControlPanel
+from vdu_controls.vdu_controller import VduController
+from vdu_controls.vdu_exceptions import VduException
 from vdu_controls.weather import WeatherQuery
 from vdu_controls.widgets import StdButton, SubWinDialog, ThemedSvgWidget, TitleButton, ThemedSvgButton, MIcon, MBox, MBtn, \
-    FasterFileDialog, PushButtonLeftJustified, ClickableSlider, LineEditAll, alter_margins, DialogSingletonMixin, ToolButton
-from vdu_controls.work_scheduler import WorkerThread, ScheduleWorker, thread_pid, SchedulerJob, SchedulerJobType, WorkException
+    FasterFileDialog, ClickableSlider, LineEditAll, alter_margins, DialogSingletonMixin, ToolButton
+from vdu_controls.work_scheduler import WorkerThread, ScheduleWorker, thread_pid, SchedulerJob, SchedulerJobType
 
 Shortcut = namedtuple('Shortcut', ['letter', 'annotated_word'])
 
@@ -96,568 +96,550 @@ def reverse_force_xwayland():
         log_info(f"Removing environment variable QT_QPA_PLATFORM")
         os.environ.pop('QT_QPA_PLATFORM')  # before the call to force_xwayland() it did not previously exist, remove it.
 
-
-class VduControllerAsyncSetter(WorkerThread):  # Used to decouple the set-vcp from the GUI
-
-    def __init__(self):
-        super().__init__(task_body=self._async_setvcp_task_body, task_finished=None, loop=True)
-        self._async_setvcp_queue: queue.Queue = queue.Queue()
-        # limit set_vcp to a sustainable interval - KDE powerdevil recommendation - 0.5s, ddcui 1.0 seconds
-        self._idle_seconds = float(os.getenv("VDU_CONTROLS_UI_IDLE_SECS", '0.5'))
-        log_info(f"env VDU_CONTROLS_UI_IDLE_SECS={self._idle_seconds}")
-
-    def _async_setvcp_task_body(self, _: WorkerThread):
-        latest_pending = {}  # Handle bursts of UI setvcp requests, filtering out repeats for the same feature.
-        while not self._async_setvcp_queue.empty():  # Keep going while there is something in the queue
-            try:
-                # print(f"{self._async_setvcp_queue.qsize()=}")
-                controller, vcp_code, value, origin = self._async_setvcp_queue.get_nowait()
-                key = (controller, vcp_code)
-                if log_debug_enabled:
-                    log_debug(f"UI discard earlier op on {controller.vdu_number=} {vcp_code=}") if key in latest_pending else None
-                latest_pending[key] = (value, origin)  # keep the latest for each controller+vcp_code.
-                self._async_setvcp_queue.task_done()
-            except queue.Empty:
-                pass
-            if latest_pending and self._async_setvcp_queue.empty():  # some setvcp requests are pending,
-                self.doze(self._idle_seconds)  # wait a bit in case more arrive - might be dragging a slider or spinning a spinner
-        if latest_pending:  # nothing more has arrived, if any setvcp requests are pending, set for real now
-            for (controller, vcp_code), (value, origin) in latest_pending.items():
-                if controller.values_cache.get(vcp_code, None) != value:
-                    log_debug(f"UI set {controller.vdu_number=} {vcp_code=} {value=} {origin}") if log_debug_enabled else None
-                    controller.set_vcp_value(vcp_code, value, origin, asynchronous_caller=True)
-                else:
-                    log_debug(f"UI nochange {controller.vdu_number=} {vcp_code=} {value=} {origin}") if log_debug_enabled else None
-        else:
-            self.doze(self._idle_seconds)
-
-    def queue_setvcp(self, controller: VduController, vcp_code: str, value: int, origin: VcpOrigin):
-        self._async_setvcp_queue.put((controller, vcp_code, value, origin))
-
-
-class VduController(QObject):
-    """
-    Holds model+controller specific to an individual VDU including a map of its capabilities. A model object in
-    MVC speak.
-
-    The model configuration can optionally be read from an INI-format config file held in $HOME/.config/vdu-control/
-
-    Capabilities are either extracted from ddcutil output or read from the INI-format files.  File read
-    capabilities are provided so that the output from "ddcutil --display N capabilities" can be corrected (because
-    it is sometimes incorrect due to sloppy implementation by manufacturers). For example, my LG monitor reports
-    two Display-Port inputs, and it only has one.
-    """
-
-    NORMAL_VDU = 0
-    IGNORE_VDU = 1
-    ASSUME_STANDARD_CONTROLS = 2
-    DISCARD_VDU = 3
-    _RANGE_PATTERN = re.compile(r'Values:\s+([0-9]+)..([0-9]+)')
-    _FEATURE_PATTERN = re.compile(r'([0-9A-F]{2})\s+[(]([^)]+)[)]\s(.*)', re.DOTALL | re.MULTILINE)
-    _LIMITED_RANGE_KEY = "%%RANGE%%"  # A key internal to vdu_controls for storing Range n..m values.
-    _FORCE_REFRESH_NAME_SUFFIX = "*refresh*"
-
-    vcp_value_changed_qtsignal = pyqtSignal(str, str, int, VcpOrigin, bool)
-    _async_setvcp_exception_qtsignal = pyqtSignal(str, int, VcpOrigin, Exception)
-
-    _async_setvcp_task: VduControllerAsyncSetter | None = None
-
-    def __init__(self, vdu_number: str, vdu_model_name: str, serial_number: str, manufacturer: str,
-                 default_config: VduControlsConfig, ddcutil: DdcutilAggregator, edit_config: Callable,
-                 vdu_exception_handler: Callable, remedy: int = 0) -> None:
-        super().__init__()
-        self.no_longer_in_use = False
-        if vdu_model_name.strip() == '':  # laptop monitors can sneak through with no model_name
-            vdu_model_name = "Unknown"
-        self.vdu_stable_id = VduStableId(proper_name(vdu_model_name, serial_number))
-        log_info(f"Initializing controls for {vdu_number=} {vdu_model_name=} {self.vdu_stable_id=}")
-        self.vdu_number = vdu_number
-        self.model_name = vdu_model_name
-        self.serial_number = serial_number
-        self.manufacturer = manufacturer
-        self.ddcutil = ddcutil
-        self.edit_config_callable = edit_config
-        self.vdu_exception_handler = vdu_exception_handler
-
-        def _handle_async_setvcp_exception(vcp_code: str, value: int, origin: VcpOrigin, e: VduException):
-            if self.vdu_exception_handler(e, True):
-                self.set_vcp_value_asynchronously(vcp_code, value, origin)
-
-        self._async_setvcp_exception_qtsignal.connect(_handle_async_setvcp_exception)
-        if VduController._async_setvcp_task is None or VduController._async_setvcp_task.isFinished():
-            VduController._async_setvcp_task = VduControllerAsyncSetter()
-            VduController._async_setvcp_task.start()
-
-        self.vdu_model_id = proper_name(vdu_model_name.strip())
-        self.capabilities_text: str | None = None
-        self.config = None
-        self.values_cache: Dict[str, int] = {}
-        self.ignore_vdu = remedy == VduController.IGNORE_VDU
-        default_sleep_multiplier: float | None = default_config.get_sleep_multiplier(fallback=None)
-        enabled_vcp_codes = default_config.get_all_enabled_vcp_codes()
-        for config_name in (self.vdu_stable_id, self.vdu_model_id):  # Allow for shared single model file (not encouraged).
-            config_path = ConfIni.get_path(config_name)
-            log_debug("checking for config file '" + config_path.as_posix() + "'") if log_debug_enabled else None
-            if os.path.isfile(config_path) and os.access(config_path, os.R_OK):
-                self.config = VduControlsConfig(config_name)
-                self.config.parse_file(config_path)
-                if default_config.is_set(ConfOpt.DEBUG_ENABLED):
-                    self.config.debug_dump()
-                enabled_vcp_codes = self.config.get_all_enabled_vcp_codes()
-                self.capabilities_text = self.config.get_capabilities_alt_text()  # cached, possibly edited, ddc capabilities
-                self.ignore_vdu = self.ignore_vdu or self.capabilities_text == '' or self.capabilities_text == IGNORE_VDU_MARKER_STR
-                if not self.ignore_vdu:
-                    if multiplier := self.config.get_sleep_multiplier(fallback=default_sleep_multiplier):
-                        self.ddcutil.set_sleep_multiplier(vdu_number, multiplier)
-                    self.ddcutil.set_vdu_specific_args(vdu_number, self.config.get_ddcutil_extra_args(fallback=None))
-                break
-        if not self.capabilities_text:
-            if remedy == VduController.DISCARD_VDU:
-                self.capabilities_text = IGNORE_VDU_MARKER_STR
-                log_info(f"Capabilities override set to ignore VDU {vdu_number=} {vdu_model_name=} {self.vdu_stable_id=}")
-            elif remedy == VduController.IGNORE_VDU:
-                self.capabilities_text = ''
-            elif remedy == VduController.ASSUME_STANDARD_CONTROLS:
-                enabled_vcp_codes = ASSUMED_CONTROLS_CONFIG_VCP_CODES
-                self.capabilities_text = ASSUMED_CONTROLS_CONFIG_TEXT
-            else:
-                self.capabilities_text = ddcutil.query_capabilities(vdu_number)
-            self.ignore_vdu = self.capabilities_text == '' or self.capabilities_text == IGNORE_VDU_MARKER_STR
-        # print(f"{self.capabilities_text}")
-        self.capabilities_supported_by_this_vdu = self._parse_capabilities(self.capabilities_text)
-        # Find those capabilities supported by this VDU AND enabled in the GUI:
-        self.enabled_capabilities = [c for c in self.capabilities_supported_by_this_vdu.values() if c.vcp_code in enabled_vcp_codes]
-        if self.config is None:  # In memory only config - in case it's needed by a future config editor
-            self.config = VduControlsConfig(self.vdu_stable_id,
-                                            default_enabled_vcp_codes=[c.vcp_code for c in self.enabled_capabilities])
-            self.config.set_capabilities_alt_text(self.capabilities_text)
-        self.config.restrict_to_actual_capabilities(self.capabilities_supported_by_this_vdu)
-        if remedy == VduController.DISCARD_VDU:
-            self.write_template_config_files()  # Persist the discard
-
-    def edit_config(self):
-        self.edit_config_callable(self.config.config_name)
-
-    def write_template_config_files(self) -> None:
-        """Write template config files to $HOME/.config/vdu_controls/"""
-        config_name = self.vdu_stable_id
-        save_config_path = ConfIni.get_path(config_name)
-        config = VduControlsConfig(config_name, default_enabled_vcp_codes=[c.vcp_code for c in self.enabled_capabilities])
-        config.set_capabilities_alt_text(self.capabilities_text if self.capabilities_text is not None else '')
-        config.write_file(save_config_path)
-        self.config = config
-
-    def get_vdu_preferred_name(self, upper: bool = False) -> str:
-        return self.config.get_vdu_preferred_name().upper() if upper else self.config.get_vdu_preferred_name()
-
-    def get_full_id(self) -> Tuple[str, str, str, str]:
-        """Return a tuple that defines this VDU: (vdu_number, manufacturer, model, serial-number)."""
-        return self.vdu_number, self.manufacturer, self.model_name, self.serial_number
-
-    def get_vcp_values(self, vcp_codes: List[str]) -> List[VcpValue]:
-        try:
-            if len(vcp_codes) == 0:
-                return []
-            # raise subprocess.SubprocessError("get_attributes")  # for testing
-            values = self.ddcutil.get_vcp_values(self.vdu_number, vcp_codes)
-            for vcp_code, vcp_value in zip(vcp_codes, values):
-                value = vcp_value.current
-                cached_value = self.values_cache.get(vcp_code, None)
-                if value != cached_value:
-                    self.values_cache[vcp_code] = value
-                    if cached_value is not None:  # Not just initialization, but an actual change...
-                        if log_debug_enabled:
-                            log_debug(
-                                f"get_vcp signals vcp_value_changed: {self.vdu_stable_id} {vcp_code=} {value} {VcpOrigin.EXTERNAL}")
-                        self.vcp_value_changed_qtsignal.emit(self.vdu_stable_id, vcp_code, value, VcpOrigin.EXTERNAL,
-                                                             self.capabilities_supported_by_this_vdu[vcp_code].causes_config_change)
-            return values
-        except (subprocess.SubprocessError, ValueError, TimeoutError, DdcutilDisplayNotFound) as e:
-            raise VduException(vdu_description=self.get_vdu_preferred_name(), vcp_code=",".join(vcp_codes), exception=e,
-                               operation="get_vcp_values") from e
-
-    def set_vcp_value(self, vcp_code: str, value: int, origin: VcpOrigin = VcpOrigin.NORMAL,
-                      asynchronous_caller: bool = False) -> None:
-        if self.no_longer_in_use:
-            log_info(f"Expired controller discards set_vcp_value({vcp_code=}, {value=}, {origin=}) {asynchronous_caller=}")
-            return
-        try:
-            # raise subprocess.SubprocessError("set_attribute")  # for testing
-            retry_on_error = vcp_code in SUPPORTED_VCP_BY_CODE and SUPPORTED_VCP_BY_CODE[vcp_code].retry_setvcp
-            self.ddcutil.set_vcp(self.vdu_number, vcp_code, value, retry_on_error=retry_on_error)
-            self.values_cache[vcp_code] = value
-            if log_debug_enabled:
-                log_debug(f"set_vcp signals vcp_value_changed: {self.vdu_stable_id} {vcp_code=} {value} {origin}")
-            self.vcp_value_changed_qtsignal.emit(self.vdu_stable_id, vcp_code, value, origin,
-                                                 self.capabilities_supported_by_this_vdu[vcp_code].causes_config_change)
-        except (subprocess.SubprocessError, ValueError, TimeoutError, DdcutilDisplayNotFound) as e:
-            vdu_exception = VduException(vdu_description=self.get_vdu_preferred_name(), vcp_code=vcp_code, exception=e,
-                                         operation="set_vcp_value")
-            if not asynchronous_caller:
-                raise vdu_exception from e
-            self._async_setvcp_exception_qtsignal.emit(vcp_code, value, origin, vdu_exception)
-
-    def set_vcp_value_asynchronously(self, vcp_code: str, value: int, origin: VcpOrigin = VcpOrigin.NORMAL) -> None:
-        # Queue the change for the queue processing thread - avoids blocking the GUI.
-        VduController._async_setvcp_task.queue_setvcp(self, vcp_code, value, origin)
-
-    def get_range_restrictions(self, vcp_code, fallback: Tuple[int, int] | None = None) -> Tuple[int, int] | None:
-        if vcp_code in self.capabilities_supported_by_this_vdu:
-            range_restriction = self.capabilities_supported_by_this_vdu[vcp_code].values  # will always be a list
-            if len(range_restriction) != 0:
-                return int(range_restriction[1]), int(range_restriction[2])
-        return fallback
-
-    def get_write_count(self):
-        return self.ddcutil.get_write_count(self.vdu_number) if self.ddcutil else 0
-
-    def _parse_capabilities(self, capabilities_text=None) -> Dict[str, VcpCapability]:
-        """Return a map of vpc capabilities keyed by vcp code."""
-
-        if capabilities_text == "Ignore VDU":
-            return {}
-
-        def _parse_values(values_str: str) -> List[str]:
-            values_list = []
-            if stripped := values_str.strip():
-                lines_list = stripped.split('\n')
-                if len(lines_list) == 1:
-                    if range_match := VduController._RANGE_PATTERN.match(lines_list[0]):
-                        values_list = [VduController._LIMITED_RANGE_KEY, range_match.group(1), range_match.group(2)]
-                    else:
-                        space_separated = lines_list[0].replace('(interpretation unavailable)', '').strip().split(' ')
-                        values_list = [(v.upper(), 'unknown ' + v) for v in space_separated[1:]]
-                else:
-                    values_list = [(key.upper(), desc.strip()) for key, desc in (v.strip().split(":", 1) for v in lines_list[1:])]
-            return values_list
-
-        feature_map = {}
-        for feature_text in capabilities_text.split(' Feature: '):
-            if feature_match := VduController._FEATURE_PATTERN.match(feature_text):
-                vcp_code = feature_match.group(1)
-                vcp_name = feature_match.group(2)
-                if requires_refresh := vcp_name.lower().endswith(VduController._FORCE_REFRESH_NAME_SUFFIX):
-                    vcp_name = vcp_name.replace(VduController._FORCE_REFRESH_NAME_SUFFIX, "")
-                values = _parse_values(feature_match.group(3))
-                # Guess type from existence or not of value list
-                if len(values) == 0:
-                    vcp_type = CONTINUOUS_TYPE
-                    if vcp_code in SUPPORTED_VCP_BY_CODE:  # Override if we know better
-                        vcp_type = SUPPORTED_VCP_BY_CODE[vcp_code].vcp_type
-                elif values[0] == VduController._LIMITED_RANGE_KEY:  # Special internal hacked config spec to specify range
-                    vcp_type = CONTINUOUS_TYPE
-                else:  # two-byte or one-byte continuous type - cannot always trust the VDU metadata on this.
-                    try:  # See whether the max is really contained within one byte:
-                        max_value = max([int(v, 16) for v, _ in values])
-                        vcp_type = COMPLEX_NON_CONTINUOUS_TYPE if max_value > 0xff else SIMPLE_NON_CONTINUOUS_TYPE
-                    except ValueError:
-                        vcp_type = COMPLEX_NON_CONTINUOUS_TYPE  # Assume two byte
-
-                capability = VcpCapability(vcp_code, vcp_name, vcp_type=vcp_type, values=values, icon_source=None,
-                                           can_transition=vcp_type == CONTINUOUS_TYPE, causes_config_change=requires_refresh)
-                feature_map[vcp_code] = capability
-        return {**{k: feature_map[k] for k in (BRIGHTNESS_VCP_CODE, CONTRAST_VCP_CODE) if k in feature_map},  # Put B&C first
-                **{k: v for k, v in feature_map.items() if k not in (BRIGHTNESS_VCP_CODE, CONTRAST_VCP_CODE)}}
-
-
-class VduException(WorkException):
-
-    def __init__(self, vdu_description=None, vcp_code=None, exception=None, operation=None) -> None:
-        super().__init__()
-        self.vdu_description = vdu_description
-        self.attr_id = vcp_code
-        self.cause = exception
-        self.operation = operation
-
-    def is_display_not_found_error(self) -> bool:
-        return self.cause is not None and isinstance(self.cause, DdcutilDisplayNotFound)
-
-    def __str__(self) -> str:
-        return f"VduException: {self.vdu_description} op={self.operation} attr={self.attr_id} {self.cause}"
-
-
-class VduControlBase(QWidget):
-    """Base GUI control for a DDC attribute."""
-
-    _refresh_ui_view_in_gui_thread_qtsignal = pyqtSignal()
-
-    def __init__(self, controller: VduController, vcp_capability: VcpCapability) -> None:
-        """Construct the slider control and initialize its values from the VDU."""
-        super().__init__()
-        self.controller = controller
-        self.vcp_capability = vcp_capability
-        self.current_value: int | None = None
-        # Using Qt signals to ensure GUI activity occurs in the GUI thread (this thread).
-        self._refresh_ui_view_in_gui_thread_qtsignal.connect(self._refresh_ui_view_task)
-        self.refresh_ui_only = False
-        self.debug = False  # Local debug switch because this is very noisy and only needed rarely.
-
-    def update_from_vdu(self, vcp_value: VcpValue):  # Used for updating from the results of get_attributes() -> List[VcpValue]
-        if self.vcp_capability.vcp_type == SIMPLE_NON_CONTINUOUS_TYPE:  # Overrides metadata value-type, enforce simple
-            self.current_value = 0x00ff & vcp_value.current  # Mask off high byte
-        else:
-            self.current_value = vcp_value.current
-        self.refresh_ui_view()
-
-    def set_value(self, new_value: int, origin: VcpOrigin = VcpOrigin.NORMAL) -> None:  # Used by controllers to alter physical VDU
-        if self.controller.values_cache[self.vcp_capability.vcp_code] != new_value:
-            self.controller.set_vcp_value(self.vcp_capability.vcp_code, new_value, origin)
-            self.current_value = new_value
-        self.refresh_ui_view()
-
-    def ui_change_vdu_attribute(self, new_value: int) -> None:  # Used by UI controls to change values
-        log_info("ui_change_vdu_attribute") if self.debug else None
-        if self.refresh_ui_only:  # Called from a GUI control when it was already responding to a vdu attribute change.
-            log_info(f"Skip change {self.refresh_ui_only=}") if self.debug else None
-            return  # Avoid repeating a setvcp by skipping the physical change
-        self.controller.set_vcp_value_asynchronously(self.vcp_capability.vcp_code, new_value, VcpOrigin.NORMAL)
-
-    def get_current_text_value(self) -> str | None:  # Return text in correct base: continuous->base10 non-continuous->base16
-        assert False, "subclass failed to implement get_current_text_value"
-
-    def refresh_ui_view_implementation(self):  # Subclasses to implement
-        assert False, "subclass failed to implement refresh_ui_view_implementation"
-
-    def refresh_ui_view(self) -> None:
-        if not is_running_in_gui_thread():
-            self._refresh_ui_view_in_gui_thread_qtsignal.emit()
-        else:
-            self._refresh_ui_view_task()
-
-    def _refresh_ui_view_task(self):
-        try:
-            self.refresh_ui_only = True  # Stop slider/widget changes from re-propagating changes to the VDU
-            self.refresh_ui_view_implementation()
-            QApplication.sendPostedEvents(self, 0)  # Flush any change events before resetting the flag
-            QApplication.processEvents()  # Force the flushed events to be processed now
-        finally:
-            self.refresh_ui_only = False
-
-
-class VduControlSlider(VduControlBase):
-    """GUI control for a DDC continuously variable attribute. A compound widget with icon, slider, and text-field."""
-
-    def __init__(self, controller: VduController, vcp_capability: VcpCapability) -> None:
-        """Construct the slider control and initialize its values from the VDU."""
-        super().__init__(controller, vcp_capability)
-        layout = QHBoxLayout()
-        self.setLayout(layout)
-        self.svg_icon: ThemedSvgWidget | None = None
-        self.setToolTip(tr(vcp_capability.name))
-        self.setToolTipDuration(TOOLTIP_DURATION_MSEC)
-        if (vcp_capability.vcp_code in SUPPORTED_VCP_BY_CODE
-                and SUPPORTED_VCP_BY_CODE[vcp_capability.vcp_code].icon_source is not None):
-            svg_icon = ThemedSvgWidget(SUPPORTED_VCP_BY_CODE[vcp_capability.vcp_code].icon_source,
-                                       native_font_height(scaled=1.8), native_font_height(scaled=1.8))
-            self.svg_icon = svg_icon
-            layout.addWidget(svg_icon)
-        else:
-            layout.addWidget(QLabel(tr(vcp_capability.name)))
-
-        self.slider = slider = ClickableSlider()
-        slider.setMinimumWidth(npx(200))
-        self.range_restriction = vcp_capability.values
-        if self.range_restriction and len(self.range_restriction) >= 2:  # Would > 2 be an error - don't worry about it
-            slider.setRange(int(self.range_restriction[1]), int(self.range_restriction[2]))
-        slider.setSingleStep(1)
-        slider.setPageStep(10)
-        slider.setTickInterval(10)
-        slider.setTickPosition(QSlider.TickPosition.TicksBelow)
-        slider.setOrientation(Qt.Orientation.Horizontal)  # type: ignore
-        slider.setTracking(False)  # Don't rewrite the ddc value too often - not sure of the implications
-        layout.addWidget(slider)
-
-        self.spinbox = QSpinBox()
-        self.spinbox.setLineEdit(LineEditAll())
-        if self.range_restriction and len(self.range_restriction) >= 2:  # >2 would be an error, but not serious
-            int_min, int_max = int(self.range_restriction[1]), int(self.range_restriction[2])
-            self.spinbox.setRange(int_min, int_max)
-            self.slider.setRange(int_min, int_max)
-
-        self.slider.setTracking(True)
-        self.slider.valueChanged.connect(self.spinbox.setValue)
-        self.spinbox.valueChanged.connect(self.slider.setValue)
-
-        self.spinbox.setValue(slider.value())
-        layout.addWidget(self.spinbox)
-
-        def _slider_changed(value: int) -> None:
-            self.current_value = value
-            self.ui_change_vdu_attribute(value)
-
-        slider.valueChanged.connect(_slider_changed)
-
-    def update_from_vdu(self, vcp_value: VcpValue):
-        if not self.range_restriction:
-            int_max = int(vcp_value.max)
-            self.spinbox.setRange(0, int_max)
-            self.slider.setRange(0, int_max)
-        super().update_from_vdu(vcp_value)
-
-    def get_current_text_value(self) -> str | None:
-        return str(self.current_value) if self.current_value is not None else None
-
-    def refresh_ui_view_implementation(self) -> None:
-        if self.current_value is not None:  # Copy the internally cached current value onto the GUI view.
-            self.slider.setValue(clamp(int(self.current_value), self.slider.minimum(), self.slider.maximum()))
-
-
-class VduControlComboBox(VduControlBase):
-    """GUI control for a DDC non-continuously variable attribute, one that has a list of choices."""
-
-    def __init__(self, controller: VduController, vcp_capability: VcpCapability) -> None:
-        super().__init__(controller, vcp_capability)
-        layout = QHBoxLayout()
-        self.setLayout(layout)
-        layout.addWidget(QLabel(self.translate_label(vcp_capability.name)))
-        self.combo_box = combo_box = QComboBox()
-        layout.addWidget(combo_box)
-        self.setToolTip(tr(vcp_capability.name))
-        self.setToolTipDuration(TOOLTIP_DURATION_MSEC)
-
-        self.keys = []
-        for value, desc in self.vcp_capability.values:
-            self.keys.append(value)
-            combo_box.addItem(self.translate_label(desc), value)
-
-        def _index_changed(_: int) -> None:
-            self.current_value = int(self.combo_box.currentData(), 16)
-            if self.validate_value() >= 0:
-                self.ui_change_vdu_attribute(self.current_value)
-
-        combo_box.currentIndexChanged.connect(_index_changed)
-
-    def translate_label(self, source: str) -> str:
-        canonical = source.lower()
-        maybe = tr(canonical)
-        result = maybe if maybe != canonical else source
-        return ' '.join(w[:1].upper() + w[1:] for w in result.split())  # Default to capitalized version of each word
-
-    def get_current_text_value(self) -> str | None:
-        return f"{self.current_value:02X}" if self.current_value is not None else None
-
-    def refresh_ui_view_implementation(self) -> None:
-        """Copy the internally cached current value onto the GUI view."""
-        value_index = self.validate_value()
-        if value_index >= 0:
-            self.combo_box.setCurrentIndex(value_index)
-
-    def validate_value(self) -> int:
-        value = self.get_current_text_value()
-        if value is None:
-            return -1
-        if value not in self.keys:
-            self.keys.append(self.current_value)
-            self.combo_box.addItem('UNKNOWN-' + value, self.current_value)
-            self.combo_box.model().item(self.combo_box.count() - 1).setEnabled(False)
-            MBox(MIcon.Critical,
-                 msg=tr("Display {vnum} {vdesc} feature {code} '({cdesc})' has an undefined value '{value}'. "
-                        "Valid values are {valid}.").format(
-                     vdesc=self.controller.get_vdu_preferred_name(), vnum=self.controller.vdu_number,
-                     code=self.vcp_capability.vcp_code, cdesc=self.vcp_capability.name,
-                     value=value, valid=self.keys),
-                 info=tr('If you want to extend the set of permitted values, you can edit the metadata '
-                         'for {} in the settings panel.  For more details see the man page concerning '
-                         'VDU/VDU-model config files.').format(self.controller.get_vdu_preferred_name())).exec()
-            return -1
-        return self.keys.index(value)
-
-
-class VduControlPanel(QWidget):
-    """Widget that contains all the controls for a single VDU (monitor/display)."""
-
-    def __init__(self, controller: VduController, vdu_exception_handler: Callable) -> None:
-        super().__init__()
-        self.controller: VduController = controller
-        layout = QVBoxLayout()
-        alter_margins(layout, top=0, bottom=0, default=self.style())
-        if int(controller.vdu_number) < 1:
-            self.title_button = TitleButton(PANEL_CONNECTED_ICON_SOURCE,
-                                            controller.get_vdu_preferred_name(),
-                                            tr("Panel {}".format(-int(controller.vdu_number))),
-                                            clicked=controller.edit_config)
-        else:
-            self.title_button = TitleButton(VDU_CONNECTED_ICON_SOURCE,
-                                            controller.get_vdu_preferred_name(),
-                                            tr("Monitor {}".format(controller.vdu_number)),
-                                            clicked=controller.edit_config)
-        layout.addWidget(self.title_button, alignment=Qt.AlignmentFlag.AlignTop)  # other params fix Qt5 theme changes
-
-        self.vcp_controls: List[VduControlBase] = []
-        self.vdu_exception_handler = vdu_exception_handler
-
-        for capability in controller.enabled_capabilities:
-            control = None
-            if capability.vcp_type == CONTINUOUS_TYPE:
-                try:
-                    control = VduControlSlider(controller, capability)
-                except ValueError as valueError:
-                    MBox(MIcon.Critical, msg=str(valueError)).exec()
-            elif capability.vcp_type in (SIMPLE_NON_CONTINUOUS_TYPE, COMPLEX_NON_CONTINUOUS_TYPE):
-                try:
-                    control = VduControlComboBox(controller, capability)
-                except ValueError as valueError:
-                    MBox(MIcon.Critical, msg=valueError.args[0],
-                         info=tr('If you want to extend the set of permitted values, see the man page concerning '
-                                 'VDU/VDU-model config files.').format(capability.vcp_code, capability.name)).exec()
-            else:
-                raise TypeError(f'No GUI support for VCP type {capability.vcp_type} for vcp_code {capability.vcp_code}')
-            if control is not None:
-                layout.addWidget(control)
-                self.vcp_controls.append(control)
-
-        line = QFrame()
-        line.setFrameShape(QFrame.Shape.HLine)
-        line.setFrameShadow(QFrame.Shadow.Sunken)
-        layout.addWidget(line)
-
-        if len(self.vcp_controls) != 0:
-            self.setLayout(layout)
-
-        self.update_stats()
-        try:
-            self.refresh_from_vdu()
-        except VduException as e:
-            self.vdu_exception_handler(e)
-
-    def get_control(self, vcp_code: str) -> VduControlBase | None:
-        return next((c for c in self.vcp_controls if c.vcp_capability.vcp_code == vcp_code), None)
-
-    def refresh_from_vdu(self) -> None:
-        """Tell the control widgets to get fresh VDU data (maybe called from a task thread, so no GUI operations here)."""
-        if values := self.controller.get_vcp_values([control.vcp_capability.vcp_code for control in self.vcp_controls]):
-            for control, value in zip(self.vcp_controls, values):
-                control.update_from_vdu(value)
-
-    def number_of_controls(self) -> int:
-        """Return the number of VDU controls.  Might be zero if initialization discovered no controllable attributes."""
-        return len(self.vcp_controls)
-
-    def is_preset_active(self, preset: Preset) -> bool:
-        vdu_section = self.controller.vdu_stable_id
-        if vdu_section == proper_name(preset.name):
-            return False  # ignore VDU initialization-presets
-        count = 0
-        preset_ini = preset.preset_ini
-        for control in self.vcp_controls:
-            if control.vcp_capability.property_name() in preset_ini[vdu_section]:
-                # Prior to version vdu_controls 1.21 we stored lower, but ddcutil expects upper
-                if control.get_current_text_value() != preset_ini[vdu_section][control.vcp_capability.property_name()].upper():
-                    return False  # immediately fail if even one value differs
-                count += 1
-        return count > 0  # true unless no values were tested.
-
-    def update_stats(self):
-        name, sid = self.controller.get_vdu_preferred_name(), self.controller.vdu_stable_id
-        title_txt = sid if id == name else f"{name}\n({sid})"
-        writes_txt = tr("Set-VCP writes: {}").format(self.controller.get_write_count())
-        if (disp_number := int(self.controller.vdu_number)) >= 0:
-            disp_numumber_txt = tr("Monitor {}").format(disp_number)
-        else:
-            disp_numumber_txt = tr("Panel {}").format(-disp_number)
-        click_txt = tr("(Click for Settings)")
-        self.title_button.setToolTip(f"{title_txt}\n{writes_txt}\n{disp_numumber_txt}\n{click_txt}")
-
+# class VduControllerAsyncSetter(WorkerThread):  # Used to decouple the set-vcp from the GUI
+#
+#     def __init__(self):
+#         super().__init__(task_body=self._async_setvcp_task_body, task_finished=None, loop=True)
+#         self._async_setvcp_queue: queue.Queue = queue.Queue()
+#         # limit set_vcp to a sustainable interval - KDE powerdevil recommendation - 0.5s, ddcui 1.0 seconds
+#         self._idle_seconds = float(os.getenv("VDU_CONTROLS_UI_IDLE_SECS", '0.5'))
+#         log_info(f"env VDU_CONTROLS_UI_IDLE_SECS={self._idle_seconds}")
+#
+#     def _async_setvcp_task_body(self, _: WorkerThread):
+#         latest_pending = {}  # Handle bursts of UI setvcp requests, filtering out repeats for the same feature.
+#         while not self._async_setvcp_queue.empty():  # Keep going while there is something in the queue
+#             try:
+#                 # print(f"{self._async_setvcp_queue.qsize()=}")
+#                 controller, vcp_code, value, origin = self._async_setvcp_queue.get_nowait()
+#                 key = (controller, vcp_code)
+#                 if log_debug_enabled:
+#                     log_debug(f"UI discard earlier op on {controller.vdu_number=} {vcp_code=}") if key in latest_pending else None
+#                 latest_pending[key] = (value, origin)  # keep the latest for each controller+vcp_code.
+#                 self._async_setvcp_queue.task_done()
+#             except queue.Empty:
+#                 pass
+#             if latest_pending and self._async_setvcp_queue.empty():  # some setvcp requests are pending,
+#                 self.doze(self._idle_seconds)  # wait a bit in case more arrive - might be dragging a slider or spinning a spinner
+#         if latest_pending:  # nothing more has arrived, if any setvcp requests are pending, set for real now
+#             for (controller, vcp_code), (value, origin) in latest_pending.items():
+#                 if controller.values_cache.get(vcp_code, None) != value:
+#                     log_debug(f"UI set {controller.vdu_number=} {vcp_code=} {value=} {origin}") if log_debug_enabled else None
+#                     controller.set_vcp_value(vcp_code, value, origin, asynchronous_caller=True)
+#                 else:
+#                     log_debug(f"UI nochange {controller.vdu_number=} {vcp_code=} {value=} {origin}") if log_debug_enabled else None
+#         else:
+#             self.doze(self._idle_seconds)
+#
+#     def queue_setvcp(self, controller: VduController, vcp_code: str, value: int, origin: VcpOrigin):
+#         self._async_setvcp_queue.put((controller, vcp_code, value, origin))
+#
+#
+# class VduController(QObject):
+#     """
+#     Holds model+controller specific to an individual VDU including a map of its capabilities. A model object in
+#     MVC speak.
+#
+#     The model configuration can optionally be read from an INI-format config file held in $HOME/.config/vdu-control/
+#
+#     Capabilities are either extracted from ddcutil output or read from the INI-format files.  File read
+#     capabilities are provided so that the output from "ddcutil --display N capabilities" can be corrected (because
+#     it is sometimes incorrect due to sloppy implementation by manufacturers). For example, my LG monitor reports
+#     two Display-Port inputs, and it only has one.
+#     """
+#
+#     NORMAL_VDU = 0
+#     IGNORE_VDU = 1
+#     ASSUME_STANDARD_CONTROLS = 2
+#     DISCARD_VDU = 3
+#     _RANGE_PATTERN = re.compile(r'Values:\s+([0-9]+)..([0-9]+)')
+#     _FEATURE_PATTERN = re.compile(r'([0-9A-F]{2})\s+[(]([^)]+)[)]\s(.*)', re.DOTALL | re.MULTILINE)
+#     _LIMITED_RANGE_KEY = "%%RANGE%%"  # A key internal to vdu_controls for storing Range n..m values.
+#     _FORCE_REFRESH_NAME_SUFFIX = "*refresh*"
+#
+#     vcp_value_changed_qtsignal = pyqtSignal(str, str, int, VcpOrigin, bool)
+#     _async_setvcp_exception_qtsignal = pyqtSignal(str, int, VcpOrigin, Exception)
+#
+#     _async_setvcp_task: VduControllerAsyncSetter | None = None
+#
+#     def __init__(self, vdu_number: str, vdu_model_name: str, serial_number: str, manufacturer: str,
+#                  default_config: VduControlsConfig, ddcutil: DdcutilAggregator, edit_config: Callable,
+#                  vdu_exception_handler: Callable, remedy: int = 0) -> None:
+#         super().__init__()
+#         self.no_longer_in_use = False
+#         if vdu_model_name.strip() == '':  # laptop monitors can sneak through with no model_name
+#             vdu_model_name = "Unknown"
+#         self.vdu_stable_id = VduStableId(proper_name(vdu_model_name, serial_number))
+#         log_info(f"Initializing controls for {vdu_number=} {vdu_model_name=} {self.vdu_stable_id=}")
+#         self.vdu_number = vdu_number
+#         self.model_name = vdu_model_name
+#         self.serial_number = serial_number
+#         self.manufacturer = manufacturer
+#         self.ddcutil = ddcutil
+#         self.edit_config_callable = edit_config
+#         self.vdu_exception_handler = vdu_exception_handler
+#
+#         def _handle_async_setvcp_exception(vcp_code: str, value: int, origin: VcpOrigin, e: VduException):
+#             if self.vdu_exception_handler(e, True):
+#                 self.set_vcp_value_asynchronously(vcp_code, value, origin)
+#
+#         self._async_setvcp_exception_qtsignal.connect(_handle_async_setvcp_exception)
+#         if VduController._async_setvcp_task is None or VduController._async_setvcp_task.isFinished():
+#             VduController._async_setvcp_task = VduControllerAsyncSetter()
+#             VduController._async_setvcp_task.start()
+#
+#         self.vdu_model_id = proper_name(vdu_model_name.strip())
+#         self.capabilities_text: str | None = None
+#         self.config = None
+#         self.values_cache: Dict[str, int] = {}
+#         self.ignore_vdu = remedy == VduController.IGNORE_VDU
+#         default_sleep_multiplier: float | None = default_config.get_sleep_multiplier(fallback=None)
+#         enabled_vcp_codes = default_config.get_all_enabled_vcp_codes()
+#         for config_name in (self.vdu_stable_id, self.vdu_model_id):  # Allow for shared single model file (not encouraged).
+#             config_path = ConfIni.get_path(config_name)
+#             log_debug("checking for config file '" + config_path.as_posix() + "'") if log_debug_enabled else None
+#             if os.path.isfile(config_path) and os.access(config_path, os.R_OK):
+#                 self.config = VduControlsConfig(config_name)
+#                 self.config.parse_file(config_path)
+#                 if default_config.is_set(ConfOpt.DEBUG_ENABLED):
+#                     self.config.debug_dump()
+#                 enabled_vcp_codes = self.config.get_all_enabled_vcp_codes()
+#                 self.capabilities_text = self.config.get_capabilities_alt_text()  # cached, possibly edited, ddc capabilities
+#                 self.ignore_vdu = self.ignore_vdu or self.capabilities_text == '' or self.capabilities_text == IGNORE_VDU_MARKER_STR
+#                 if not self.ignore_vdu:
+#                     if multiplier := self.config.get_sleep_multiplier(fallback=default_sleep_multiplier):
+#                         self.ddcutil.set_sleep_multiplier(vdu_number, multiplier)
+#                     self.ddcutil.set_vdu_specific_args(vdu_number, self.config.get_ddcutil_extra_args(fallback=None))
+#                 break
+#         if not self.capabilities_text:
+#             if remedy == VduController.DISCARD_VDU:
+#                 self.capabilities_text = IGNORE_VDU_MARKER_STR
+#                 log_info(f"Capabilities override set to ignore VDU {vdu_number=} {vdu_model_name=} {self.vdu_stable_id=}")
+#             elif remedy == VduController.IGNORE_VDU:
+#                 self.capabilities_text = ''
+#             elif remedy == VduController.ASSUME_STANDARD_CONTROLS:
+#                 enabled_vcp_codes = ASSUMED_CONTROLS_CONFIG_VCP_CODES
+#                 self.capabilities_text = ASSUMED_CONTROLS_CONFIG_TEXT
+#             else:
+#                 self.capabilities_text = ddcutil.query_capabilities(vdu_number)
+#             self.ignore_vdu = self.capabilities_text == '' or self.capabilities_text == IGNORE_VDU_MARKER_STR
+#         # print(f"{self.capabilities_text}")
+#         self.capabilities_supported_by_this_vdu = self._parse_capabilities(self.capabilities_text)
+#         # Find those capabilities supported by this VDU AND enabled in the GUI:
+#         self.enabled_capabilities = [c for c in self.capabilities_supported_by_this_vdu.values() if c.vcp_code in enabled_vcp_codes]
+#         if self.config is None:  # In memory only config - in case it's needed by a future config editor
+#             self.config = VduControlsConfig(self.vdu_stable_id,
+#                                             default_enabled_vcp_codes=[c.vcp_code for c in self.enabled_capabilities])
+#             self.config.set_capabilities_alt_text(self.capabilities_text)
+#         self.config.restrict_to_actual_capabilities(self.capabilities_supported_by_this_vdu)
+#         if remedy == VduController.DISCARD_VDU:
+#             self.write_template_config_files()  # Persist the discard
+#
+#     def edit_config(self):
+#         self.edit_config_callable(self.config.config_name)
+#
+#     def write_template_config_files(self) -> None:
+#         """Write template config files to $HOME/.config/vdu_controls/"""
+#         config_name = self.vdu_stable_id
+#         save_config_path = ConfIni.get_path(config_name)
+#         config = VduControlsConfig(config_name, default_enabled_vcp_codes=[c.vcp_code for c in self.enabled_capabilities])
+#         config.set_capabilities_alt_text(self.capabilities_text if self.capabilities_text is not None else '')
+#         config.write_file(save_config_path)
+#         self.config = config
+#
+#     def get_vdu_preferred_name(self, upper: bool = False) -> str:
+#         return self.config.get_vdu_preferred_name().upper() if upper else self.config.get_vdu_preferred_name()
+#
+#     def get_full_id(self) -> Tuple[str, str, str, str]:
+#         """Return a tuple that defines this VDU: (vdu_number, manufacturer, model, serial-number)."""
+#         return self.vdu_number, self.manufacturer, self.model_name, self.serial_number
+#
+#     def get_vcp_values(self, vcp_codes: List[str]) -> List[VcpValue]:
+#         try:
+#             if len(vcp_codes) == 0:
+#                 return []
+#             # raise subprocess.SubprocessError("get_attributes")  # for testing
+#             values = self.ddcutil.get_vcp_values(self.vdu_number, vcp_codes)
+#             for vcp_code, vcp_value in zip(vcp_codes, values):
+#                 value = vcp_value.current
+#                 cached_value = self.values_cache.get(vcp_code, None)
+#                 if value != cached_value:
+#                     self.values_cache[vcp_code] = value
+#                     if cached_value is not None:  # Not just initialization, but an actual change...
+#                         if log_debug_enabled:
+#                             log_debug(
+#                                 f"get_vcp signals vcp_value_changed: {self.vdu_stable_id} {vcp_code=} {value} {VcpOrigin.EXTERNAL}")
+#                         self.vcp_value_changed_qtsignal.emit(self.vdu_stable_id, vcp_code, value, VcpOrigin.EXTERNAL,
+#                                                              self.capabilities_supported_by_this_vdu[vcp_code].causes_config_change)
+#             return values
+#         except (subprocess.SubprocessError, ValueError, TimeoutError, DdcutilDisplayNotFound) as e:
+#             raise VduException(vdu_description=self.get_vdu_preferred_name(), vcp_code=",".join(vcp_codes), exception=e,
+#                                operation="get_vcp_values") from e
+#
+#     def set_vcp_value(self, vcp_code: str, value: int, origin: VcpOrigin = VcpOrigin.NORMAL,
+#                       asynchronous_caller: bool = False) -> None:
+#         if self.no_longer_in_use:
+#             log_info(f"Expired controller discards set_vcp_value({vcp_code=}, {value=}, {origin=}) {asynchronous_caller=}")
+#             return
+#         try:
+#             # raise subprocess.SubprocessError("set_attribute")  # for testing
+#             retry_on_error = vcp_code in SUPPORTED_VCP_BY_CODE and SUPPORTED_VCP_BY_CODE[vcp_code].retry_setvcp
+#             self.ddcutil.set_vcp(self.vdu_number, vcp_code, value, retry_on_error=retry_on_error)
+#             self.values_cache[vcp_code] = value
+#             if log_debug_enabled:
+#                 log_debug(f"set_vcp signals vcp_value_changed: {self.vdu_stable_id} {vcp_code=} {value} {origin}")
+#             self.vcp_value_changed_qtsignal.emit(self.vdu_stable_id, vcp_code, value, origin,
+#                                                  self.capabilities_supported_by_this_vdu[vcp_code].causes_config_change)
+#         except (subprocess.SubprocessError, ValueError, TimeoutError, DdcutilDisplayNotFound) as e:
+#             vdu_exception = VduException(vdu_description=self.get_vdu_preferred_name(), vcp_code=vcp_code, exception=e,
+#                                          operation="set_vcp_value")
+#             if not asynchronous_caller:
+#                 raise vdu_exception from e
+#             self._async_setvcp_exception_qtsignal.emit(vcp_code, value, origin, vdu_exception)
+#
+#     def set_vcp_value_asynchronously(self, vcp_code: str, value: int, origin: VcpOrigin = VcpOrigin.NORMAL) -> None:
+#         # Queue the change for the queue processing thread - avoids blocking the GUI.
+#         VduController._async_setvcp_task.queue_setvcp(self, vcp_code, value, origin)
+#
+#     def get_range_restrictions(self, vcp_code, fallback: Tuple[int, int] | None = None) -> Tuple[int, int] | None:
+#         if vcp_code in self.capabilities_supported_by_this_vdu:
+#             range_restriction = self.capabilities_supported_by_this_vdu[vcp_code].values  # will always be a list
+#             if len(range_restriction) != 0:
+#                 return int(range_restriction[1]), int(range_restriction[2])
+#         return fallback
+#
+#     def get_write_count(self):
+#         return self.ddcutil.get_write_count(self.vdu_number) if self.ddcutil else 0
+#
+#     def _parse_capabilities(self, capabilities_text=None) -> Dict[str, VcpCapability]:
+#         """Return a map of vpc capabilities keyed by vcp code."""
+#
+#         if capabilities_text == "Ignore VDU":
+#             return {}
+#
+#         def _parse_values(values_str: str) -> List[str]:
+#             values_list = []
+#             if stripped := values_str.strip():
+#                 lines_list = stripped.split('\n')
+#                 if len(lines_list) == 1:
+#                     if range_match := VduController._RANGE_PATTERN.match(lines_list[0]):
+#                         values_list = [VduController._LIMITED_RANGE_KEY, range_match.group(1), range_match.group(2)]
+#                     else:
+#                         space_separated = lines_list[0].replace('(interpretation unavailable)', '').strip().split(' ')
+#                         values_list = [(v.upper(), 'unknown ' + v) for v in space_separated[1:]]
+#                 else:
+#                     values_list = [(key.upper(), desc.strip()) for key, desc in (v.strip().split(":", 1) for v in lines_list[1:])]
+#             return values_list
+#
+#         feature_map = {}
+#         for feature_text in capabilities_text.split(' Feature: '):
+#             if feature_match := VduController._FEATURE_PATTERN.match(feature_text):
+#                 vcp_code = feature_match.group(1)
+#                 vcp_name = feature_match.group(2)
+#                 if requires_refresh := vcp_name.lower().endswith(VduController._FORCE_REFRESH_NAME_SUFFIX):
+#                     vcp_name = vcp_name.replace(VduController._FORCE_REFRESH_NAME_SUFFIX, "")
+#                 values = _parse_values(feature_match.group(3))
+#                 # Guess type from existence or not of value list
+#                 if len(values) == 0:
+#                     vcp_type = CONTINUOUS_TYPE
+#                     if vcp_code in SUPPORTED_VCP_BY_CODE:  # Override if we know better
+#                         vcp_type = SUPPORTED_VCP_BY_CODE[vcp_code].vcp_type
+#                 elif values[0] == VduController._LIMITED_RANGE_KEY:  # Special internal hacked config spec to specify range
+#                     vcp_type = CONTINUOUS_TYPE
+#                 else:  # two-byte or one-byte continuous type - cannot always trust the VDU metadata on this.
+#                     try:  # See whether the max is really contained within one byte:
+#                         max_value = max([int(v, 16) for v, _ in values])
+#                         vcp_type = COMPLEX_NON_CONTINUOUS_TYPE if max_value > 0xff else SIMPLE_NON_CONTINUOUS_TYPE
+#                     except ValueError:
+#                         vcp_type = COMPLEX_NON_CONTINUOUS_TYPE  # Assume two byte
+#
+#                 capability = VcpCapability(vcp_code, vcp_name, vcp_type=vcp_type, values=values, icon_source=None,
+#                                            can_transition=vcp_type == CONTINUOUS_TYPE, causes_config_change=requires_refresh)
+#                 feature_map[vcp_code] = capability
+#         return {**{k: feature_map[k] for k in (BRIGHTNESS_VCP_CODE, CONTRAST_VCP_CODE) if k in feature_map},  # Put B&C first
+#                 **{k: v for k, v in feature_map.items() if k not in (BRIGHTNESS_VCP_CODE, CONTRAST_VCP_CODE)}}
+
+# class VduControlBase(QWidget):
+#     """Base GUI control for a DDC attribute."""
+#
+#     _refresh_ui_view_in_gui_thread_qtsignal = pyqtSignal()
+#
+#     def __init__(self, controller: VduController, vcp_capability: VcpCapability) -> None:
+#         """Construct the slider control and initialize its values from the VDU."""
+#         super().__init__()
+#         self.controller = controller
+#         self.vcp_capability = vcp_capability
+#         self.current_value: int | None = None
+#         # Using Qt signals to ensure GUI activity occurs in the GUI thread (this thread).
+#         self._refresh_ui_view_in_gui_thread_qtsignal.connect(self._refresh_ui_view_task)
+#         self.refresh_ui_only = False
+#         self.debug = False  # Local debug switch because this is very noisy and only needed rarely.
+#
+#     def update_from_vdu(self, vcp_value: VcpValue):  # Used for updating from the results of get_attributes() -> List[VcpValue]
+#         if self.vcp_capability.vcp_type == SIMPLE_NON_CONTINUOUS_TYPE:  # Overrides metadata value-type, enforce simple
+#             self.current_value = 0x00ff & vcp_value.current  # Mask off high byte
+#         else:
+#             self.current_value = vcp_value.current
+#         self.refresh_ui_view()
+#
+#     def set_value(self, new_value: int, origin: VcpOrigin = VcpOrigin.NORMAL) -> None:  # Used by controllers to alter physical VDU
+#         if self.controller.values_cache[self.vcp_capability.vcp_code] != new_value:
+#             self.controller.set_vcp_value(self.vcp_capability.vcp_code, new_value, origin)
+#             self.current_value = new_value
+#         self.refresh_ui_view()
+#
+#     def ui_change_vdu_attribute(self, new_value: int) -> None:  # Used by UI controls to change values
+#         log_info("ui_change_vdu_attribute") if self.debug else None
+#         if self.refresh_ui_only:  # Called from a GUI control when it was already responding to a vdu attribute change.
+#             log_info(f"Skip change {self.refresh_ui_only=}") if self.debug else None
+#             return  # Avoid repeating a setvcp by skipping the physical change
+#         self.controller.set_vcp_value_asynchronously(self.vcp_capability.vcp_code, new_value, VcpOrigin.NORMAL)
+#
+#     def get_current_text_value(self) -> str | None:  # Return text in correct base: continuous->base10 non-continuous->base16
+#         assert False, "subclass failed to implement get_current_text_value"
+#
+#     def refresh_ui_view_implementation(self):  # Subclasses to implement
+#         assert False, "subclass failed to implement refresh_ui_view_implementation"
+#
+#     def refresh_ui_view(self) -> None:
+#         if not is_running_in_gui_thread():
+#             self._refresh_ui_view_in_gui_thread_qtsignal.emit()
+#         else:
+#             self._refresh_ui_view_task()
+#
+#     def _refresh_ui_view_task(self):
+#         try:
+#             self.refresh_ui_only = True  # Stop slider/widget changes from re-propagating changes to the VDU
+#             self.refresh_ui_view_implementation()
+#             QApplication.sendPostedEvents(self, 0)  # Flush any change events before resetting the flag
+#             QApplication.processEvents()  # Force the flushed events to be processed now
+#         finally:
+#             self.refresh_ui_only = False
+#
+#
+# class VduControlSlider(VduControlBase):
+#     """GUI control for a DDC continuously variable attribute. A compound widget with icon, slider, and text-field."""
+#
+#     def __init__(self, controller: VduController, vcp_capability: VcpCapability) -> None:
+#         """Construct the slider control and initialize its values from the VDU."""
+#         super().__init__(controller, vcp_capability)
+#         layout = QHBoxLayout()
+#         self.setLayout(layout)
+#         self.svg_icon: ThemedSvgWidget | None = None
+#         self.setToolTip(tr(vcp_capability.name))
+#         self.setToolTipDuration(TOOLTIP_DURATION_MSEC)
+#         if (vcp_capability.vcp_code in SUPPORTED_VCP_BY_CODE
+#                 and SUPPORTED_VCP_BY_CODE[vcp_capability.vcp_code].icon_source is not None):
+#             svg_icon = ThemedSvgWidget(SUPPORTED_VCP_BY_CODE[vcp_capability.vcp_code].icon_source,
+#                                        native_font_height(scaled=1.8), native_font_height(scaled=1.8))
+#             self.svg_icon = svg_icon
+#             layout.addWidget(svg_icon)
+#         else:
+#             layout.addWidget(QLabel(tr(vcp_capability.name)))
+#
+#         self.slider = slider = ClickableSlider()
+#         slider.setMinimumWidth(npx(200))
+#         self.range_restriction = vcp_capability.values
+#         if self.range_restriction and len(self.range_restriction) >= 2:  # Would > 2 be an error - don't worry about it
+#             slider.setRange(int(self.range_restriction[1]), int(self.range_restriction[2]))
+#         slider.setSingleStep(1)
+#         slider.setPageStep(10)
+#         slider.setTickInterval(10)
+#         slider.setTickPosition(QSlider.TickPosition.TicksBelow)
+#         slider.setOrientation(Qt.Orientation.Horizontal)  # type: ignore
+#         slider.setTracking(False)  # Don't rewrite the ddc value too often - not sure of the implications
+#         layout.addWidget(slider)
+#
+#         self.spinbox = QSpinBox()
+#         self.spinbox.setLineEdit(LineEditAll())
+#         if self.range_restriction and len(self.range_restriction) >= 2:  # >2 would be an error, but not serious
+#             int_min, int_max = int(self.range_restriction[1]), int(self.range_restriction[2])
+#             self.spinbox.setRange(int_min, int_max)
+#             self.slider.setRange(int_min, int_max)
+#
+#         self.slider.setTracking(True)
+#         self.slider.valueChanged.connect(self.spinbox.setValue)
+#         self.spinbox.valueChanged.connect(self.slider.setValue)
+#
+#         self.spinbox.setValue(slider.value())
+#         layout.addWidget(self.spinbox)
+#
+#         def _slider_changed(value: int) -> None:
+#             self.current_value = value
+#             self.ui_change_vdu_attribute(value)
+#
+#         slider.valueChanged.connect(_slider_changed)
+#
+#     def update_from_vdu(self, vcp_value: VcpValue):
+#         if not self.range_restriction:
+#             int_max = int(vcp_value.max)
+#             self.spinbox.setRange(0, int_max)
+#             self.slider.setRange(0, int_max)
+#         super().update_from_vdu(vcp_value)
+#
+#     def get_current_text_value(self) -> str | None:
+#         return str(self.current_value) if self.current_value is not None else None
+#
+#     def refresh_ui_view_implementation(self) -> None:
+#         if self.current_value is not None:  # Copy the internally cached current value onto the GUI view.
+#             self.slider.setValue(clamp(int(self.current_value), self.slider.minimum(), self.slider.maximum()))
+#
+#
+# class VduControlComboBox(VduControlBase):
+#     """GUI control for a DDC non-continuously variable attribute, one that has a list of choices."""
+#
+#     def __init__(self, controller: VduController, vcp_capability: VcpCapability) -> None:
+#         super().__init__(controller, vcp_capability)
+#         layout = QHBoxLayout()
+#         self.setLayout(layout)
+#         layout.addWidget(QLabel(self.translate_label(vcp_capability.name)))
+#         self.combo_box = combo_box = QComboBox()
+#         layout.addWidget(combo_box)
+#         self.setToolTip(tr(vcp_capability.name))
+#         self.setToolTipDuration(TOOLTIP_DURATION_MSEC)
+#
+#         self.keys = []
+#         for value, desc in self.vcp_capability.values:
+#             self.keys.append(value)
+#             combo_box.addItem(self.translate_label(desc), value)
+#
+#         def _index_changed(_: int) -> None:
+#             self.current_value = int(self.combo_box.currentData(), 16)
+#             if self.validate_value() >= 0:
+#                 self.ui_change_vdu_attribute(self.current_value)
+#
+#         combo_box.currentIndexChanged.connect(_index_changed)
+#
+#     def translate_label(self, source: str) -> str:
+#         canonical = source.lower()
+#         maybe = tr(canonical)
+#         result = maybe if maybe != canonical else source
+#         return ' '.join(w[:1].upper() + w[1:] for w in result.split())  # Default to capitalized version of each word
+#
+#     def get_current_text_value(self) -> str | None:
+#         return f"{self.current_value:02X}" if self.current_value is not None else None
+#
+#     def refresh_ui_view_implementation(self) -> None:
+#         """Copy the internally cached current value onto the GUI view."""
+#         value_index = self.validate_value()
+#         if value_index >= 0:
+#             self.combo_box.setCurrentIndex(value_index)
+#
+#     def validate_value(self) -> int:
+#         value = self.get_current_text_value()
+#         if value is None:
+#             return -1
+#         if value not in self.keys:
+#             self.keys.append(self.current_value)
+#             self.combo_box.addItem('UNKNOWN-' + value, self.current_value)
+#             self.combo_box.model().item(self.combo_box.count() - 1).setEnabled(False)
+#             MBox(MIcon.Critical,
+#                  msg=tr("Display {vnum} {vdesc} feature {code} '({cdesc})' has an undefined value '{value}'. "
+#                         "Valid values are {valid}.").format(
+#                      vdesc=self.controller.get_vdu_preferred_name(), vnum=self.controller.vdu_number,
+#                      code=self.vcp_capability.vcp_code, cdesc=self.vcp_capability.name,
+#                      value=value, valid=self.keys),
+#                  info=tr('If you want to extend the set of permitted values, you can edit the metadata '
+#                          'for {} in the settings panel.  For more details see the man page concerning '
+#                          'VDU/VDU-model config files.').format(self.controller.get_vdu_preferred_name())).exec()
+#             return -1
+#         return self.keys.index(value)
+#
+#
+# class VduControlPanel(QWidget):
+#     """Widget that contains all the controls for a single VDU (monitor/display)."""
+#
+#     def __init__(self, controller: VduController, vdu_exception_handler: Callable) -> None:
+#         super().__init__()
+#         self.controller: VduController = controller
+#         layout = QVBoxLayout()
+#         alter_margins(layout, top=0, bottom=0, default=self.style())
+#         if int(controller.vdu_number) < 1:
+#             self.title_button = TitleButton(PANEL_CONNECTED_ICON_SOURCE,
+#                                             controller.get_vdu_preferred_name(),
+#                                             tr("Panel {}".format(-int(controller.vdu_number))),
+#                                             clicked=controller.edit_config)
+#         else:
+#             self.title_button = TitleButton(VDU_CONNECTED_ICON_SOURCE,
+#                                             controller.get_vdu_preferred_name(),
+#                                             tr("Monitor {}".format(controller.vdu_number)),
+#                                             clicked=controller.edit_config)
+#         layout.addWidget(self.title_button, alignment=Qt.AlignmentFlag.AlignTop)  # other params fix Qt5 theme changes
+#
+#         self.vcp_controls: List[VduControlBase] = []
+#         self.vdu_exception_handler = vdu_exception_handler
+#
+#         for capability in controller.enabled_capabilities:
+#             control = None
+#             if capability.vcp_type == CONTINUOUS_TYPE:
+#                 try:
+#                     control = VduControlSlider(controller, capability)
+#                 except ValueError as valueError:
+#                     MBox(MIcon.Critical, msg=str(valueError)).exec()
+#             elif capability.vcp_type in (SIMPLE_NON_CONTINUOUS_TYPE, COMPLEX_NON_CONTINUOUS_TYPE):
+#                 try:
+#                     control = VduControlComboBox(controller, capability)
+#                 except ValueError as valueError:
+#                     MBox(MIcon.Critical, msg=valueError.args[0],
+#                          info=tr('If you want to extend the set of permitted values, see the man page concerning '
+#                                  'VDU/VDU-model config files.').format(capability.vcp_code, capability.name)).exec()
+#             else:
+#                 raise TypeError(f'No GUI support for VCP type {capability.vcp_type} for vcp_code {capability.vcp_code}')
+#             if control is not None:
+#                 layout.addWidget(control)
+#                 self.vcp_controls.append(control)
+#
+#         line = QFrame()
+#         line.setFrameShape(QFrame.Shape.HLine)
+#         line.setFrameShadow(QFrame.Shadow.Sunken)
+#         layout.addWidget(line)
+#
+#         if len(self.vcp_controls) != 0:
+#             self.setLayout(layout)
+#
+#         self.update_stats()
+#         try:
+#             self.refresh_from_vdu()
+#         except VduException as e:
+#             self.vdu_exception_handler(e)
+#
+#     def get_control(self, vcp_code: str) -> VduControlBase | None:
+#         return next((c for c in self.vcp_controls if c.vcp_capability.vcp_code == vcp_code), None)
+#
+#     def refresh_from_vdu(self) -> None:
+#         """Tell the control widgets to get fresh VDU data (maybe called from a task thread, so no GUI operations here)."""
+#         if values := self.controller.get_vcp_values([control.vcp_capability.vcp_code for control in self.vcp_controls]):
+#             for control, value in zip(self.vcp_controls, values):
+#                 control.update_from_vdu(value)
+#
+#     def number_of_controls(self) -> int:
+#         """Return the number of VDU controls.  Might be zero if initialization discovered no controllable attributes."""
+#         return len(self.vcp_controls)
+#
+#     def is_preset_active(self, preset: Preset) -> bool:
+#         vdu_section = self.controller.vdu_stable_id
+#         if vdu_section == proper_name(preset.name):
+#             return False  # ignore VDU initialization-presets
+#         count = 0
+#         preset_ini = preset.preset_ini
+#         for control in self.vcp_controls:
+#             if control.vcp_capability.property_name() in preset_ini[vdu_section]:
+#                 # Prior to version vdu_controls 1.21 we stored lower, but ddcutil expects upper
+#                 if control.get_current_text_value() != preset_ini[vdu_section][control.vcp_capability.property_name()].upper():
+#                     return False  # immediately fail if even one value differs
+#                 count += 1
+#         return count > 0  # true unless no values were tested.
+#
+#     def update_stats(self):
+#         name, sid = self.controller.get_vdu_preferred_name(), self.controller.vdu_stable_id
+#         title_txt = sid if id == name else f"{name}\n({sid})"
+#         writes_txt = tr("Set-VCP writes: {}").format(self.controller.get_write_count())
+#         if (disp_number := int(self.controller.vdu_number)) >= 0:
+#             disp_numumber_txt = tr("Monitor {}").format(disp_number)
+#         else:
+#             disp_numumber_txt = tr("Panel {}").format(-disp_number)
+#         click_txt = tr("(Click for Settings)")
+#         self.title_button.setToolTip(f"{title_txt}\n{writes_txt}\n{disp_numumber_txt}\n{click_txt}")
+#
 
 class ContextMenu(QMenu):
     PRESET_NAME_PROP = 'preset_name'
